requirements.txt
# Core
pdfplumber>=0.11.0
networkx>=3.3
rich>=13.7.0

# LLM providers — install at least one
anthropic>=0.30.0          # for Claude (Anthropic)
google-generativeai>=0.8.0 # for Gemini (Google) — optional

# Dev / testing
pytest>=8.0.0
pytest-cov>=5.0.0